hospital_mgmt/requirements.txt
asgiref==3.10.0
Django==5.2.7
mysql-connector-python==9.4.0
mysqlclient==2.2.7
pillow==12.0.0
sqlparse==0.5.3
